Changed what happens when you add item twice
Now when you add an item to the cart that is arleady in
that cart, the quantity will be increased by one on both the
front end and back end.  Created a new modal in index for this
and also a new function to handle this event.  Also changed error
message for adding to cart and got rid of modal that stopped you from
adding an item twice